backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
pydantic
redis
openai
requests
python-multipart
